feat: 上游同步批次 4 — 全量功能同步 + 测试补全 (146 files, +27515/-2161)
功能同步:
- cmd: config/import/setup/skill 命令增强
- agent: curator/fallback/insights/ratelimit/think_scrubber/trajectory 更新
- gateway: 全平台适配器重构 (18 个平台)、delivery/pairing/runner/session 优化
- hooks: executor 重写 + 测试扩展
- cron: delivery/executor/jobs/scheduler 更新
- tool: browser/camofox/checkpoint/file/patch/terminal/web 等 30+ 工具更新
- mcp: adapter/client/oauth/protocol/server 协议升级
- plugin/memory/sandbox/credential/batch 各模块功能同步
- i18n: locale 翻译更新
- telemetry: sink 更新

测试补全:
- agent: 13 个新测试文件 (auxiliary/conversation_unit/curator/fallback/insights/
  options/ratelimit/redact/retry/router/shell_hooks/title/trajectory)
- context: 5 个新测试文件 (compressor_prune/summarize/hints/prompt_cache/references)
- gateway: 9 个新测试文件 (cache/channel_dir/chunking/delivery/hooks/pairing/runner/stream)
- llm: 7 个新测试文件 (anthropic_thinking/caching/client/copilot_acp/estimate/
  eventstream/stream)
- credential: pool_test
- state: state_test

清理:
- 删除过期的 docs/REWRITE_PLAN.md

+++ internal/tool/yuanbao.go
@@ -464,8 +464,8 @@ func (t *YBSearchStickerTool) Execute(ctx context.Context, args map[string]any) 
 	query, _ := args["query"].(string)
 
 	limit := 10
-	if l, ok := args["limit"].(int); ok && l > 0 {
-		limit = l
+	if l, ok := args["limit"].(float64); ok && l > 0 {
+		limit = int(l)
 	}
 	if limit > 50 {
 		limit = 50
